Cleanup `pwnlib.lexer` exports and imports (#2611)
* Cleanup `pwnlib.lexer` exports and imports

The exported lexers don't really exist. It has been this way since this code was first included in the repo, so I'd assume the intention was to have aliases for the same lexer since the disassembly looks similar between all those tools.

Many imports are unused too, so remove them.

Fixes #1309

* Update CHANGELOG