ROP: Remove stdout and argv workaround in ROPgadget invocation (#2639)
* ROP: Remove stdout and argv workaround in ROPgadget invocation

ROPgadget supports usage as a library since v5.3 in 2015.
The hacks are no longer needed.

This also fixes running multiple exploits in parallel in the same process. They could race and output was lost.

* Update CHANGELOG

* Fix race while creating rop-cache folder

There was a TOCTOU when two ROP classes were interleaved just wrong in a multi-threaded script.